bot/requirements.txt
# Core Trading Libraries
MetaTrader5==5.0.45
pandas==2.1.4
numpy==1.26.2

# Machine Learning
stable-baselines3==2.2.1
torch==2.1.2
gymnasium==0.29.1
scikit-learn==1.3.2

# Technical Analysis
ta-lib==0.4.28

# Database
psycopg2-binary==2.9.9

# AWS Integration
boto3==1.34.14
botocore==1.34.14

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
